Added a random shuffle of arrays before creating a TFDataset.
